+++ packages/common/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 
 ### Added
+
+- Helpers for calculating energy cost for a transaction and microCCD cost from energy cost:
+  - `getEnergyCost`
+  - `getExchangeRate`
+  - `convertEnergyToMicroCcd`
 - Utility functions for extracting information from `BlockItemSummary`:
     - `isInitContractSummary`
     - `isUpdateContractSummary`