+++ packages/common/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 7.0.0 2023-05-15 
 
 ### Breaking changes
+
 - Updated `blockInfo` so that the `bakerId` field is optional, since it will be undefined for genesis blocks.
 - `waitForTransactionFinalization` now returns a `BlockItemSummaryInBlock`
 - Added missing version return type in `getModuleSchema`. It now returns an object containing the schema source and version.
@@ -30,6 +31,11 @@
     - `findEarliestFinalized`
     - `findInstanceCreation`
     - `findFirstFinalizedBlockNoLaterThan`
+- Extended HdWallet with support for verifiable credential key deriviation.
+
+### Changed
+
+- Bumped @concordium/rust-bindings to 0.12.0. (Adds key derivation for verifiable credentials)
 
 ## 6.5.0 2023-5-03